requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
chromadb>=0.4.15
pypdf>=3.15.1
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2